feat: refactor multislect into own component
